requirements.txt
fastapi
uvicorn
streamlit
sseclient-py
requests
python-dotenv
langchain
langchain-openai
langgraph
faiss-cpu 
langchain-community